Move SnowflakePredicateBuilder's code in SnowflakeExpressionBuilder (#257)
And make `Predicate` a simple marker trait (which doesn't extend
`Expression` anymore). This is because some expressions such as literals
`True` and `False` cannot be made predicates (although they logically
are)
